Fix cookie.expires date assignment
I caught an error while creating a new session with this library. I tried downloading the code and log it to see what could be the cause, and it showed to me that the cookie generated here always had a 'Invalid Date' value, resulting in a MySQL error after creating a new user.

Adding one more validation (!NaN(data.cookie.expires)) fixed this issue for me. Please, check if someone else have this problem and if this patch fixes it.