Remove unused dependency: `deep-equal`
Aside `assert.deepEqual()` being native and faster, this dependency just
isn’t being used. I did _not_ regenerate the `package-lock.json` though
because it seems there are dependabot and an `npm audit` errro that
should be addressed—along with the version of package ≠ version in the
lockfile—so I thought that part would be best left to the maintainer.