fix/findCreateFind should always use master
When using sequelize with read replica instance we can race condition on the findCreateFind query.

I just starting testing the read replica configuration and encountered the following bug when creating a session :
```Cannot read properties of null (reading 'dataValues')```
This can happen because while the session was written in the master db the following find query is made in the read replica which is not updated yet due to latency.

Adding ``` useMaster: true``` option to the findCreateFind query should force the find query to be made against the master db thus avoiding the bug.
More information on sequelize read replication : https://sequelize.org/docs/v6/other-topics/read-replication/

I am not able to replicate the issue with the exemple repository since it requires a specific databse setup. If you need more information I would be happy to help.